+++ internal/pkg/key/key.go
@@ -19,6 +19,7 @@ const (
 	TeleportBotSecretName           = "identity-output"
 	TeleportBotNamespace            = "giantswarm"
 	TeleportBotAppName              = "teleport-tbot"
+	TbotOutputsConfigmapName        = "teleport-tbot-outputs"
 	TeleportAppTokenValidity        = 720 * time.Hour
 	TeleportKubeTokenValidity       = 720 * time.Hour
 	TeleportNodeTokenValidity       = 720 * time.Hour